requirements.txt
cmake>=3.13.2
plyfile
tensorboardX
pyyaml
matplotlib
opencv-python
trimesh>=2.35.39,<2.35.40
networkx>=2.2,<2.3
scipy
trimesh